Refactor page folder
