+++ web/src/app/pages/Alumni/AlumniSignup.page.tsx
@@ -1,6 +1,6 @@
-import SignupForm from '../components/AuthForms/SignupForm';
-import { Role } from '../type/role';
-import classes from '../styles/LoginPage.module.css';
+import SignupForm from '../../components/AuthForms/SignupForm';
+import { Role } from '../../type/role';
+import classes from '../../styles/LoginPage.module.css';
 import { Box } from '@mantine/core';
 
 export function AlumniSignUp() {
