+++ web/src/app/pages/Sponsor/SponsorProfile.page.tsx
@@ -1,10 +1,10 @@
 import { Card, Avatar, Text, Box, Title, Button, Grid, Flex, Loader } from '@mantine/core';
-import classes from '../styles/SponsorProfile.module.css';
+import classes from '../../styles/SponsorProfile.module.css';
 import { useEffect, useState } from 'react';
 import { IconPlus } from '@tabler/icons-react';
-import { JobCarousel } from '../components/JobCardCarousel/JobCarousel';
-import { JobCardProps } from '../components/JobCardCarousel/JobCard';
-import { UserType } from '../features/user/userSlice';
+import { JobCarousel } from '../../components/JobCardCarousel/JobCarousel';
+import { JobCardProps } from '../../components/JobCardCarousel/JobCard';
+import { UserType } from '../../features/user/userSlice';
 
 export function SponsorProfile() {
   // UseState for future modal implementation
